backend/pyproject.toml
[project]
name = "nevertrtfm"
version = "0.1.0"
description = "Upload any report. Get a TikTok-style video. Ask questions live."
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "google-adk>=0.4.0",
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-texttospeech>=2.17.0",
    "google-cloud-storage>=2.18.0",
    "websockets>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
